backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
youtube-transcript-api==0.6.1
python-dotenv==1.0.0
google-generativeai==0.3.2 